Add pdb2structs.py and a Windows 7 preload
pdb2structs.py accepts a .pdb file (Windows debugging information) and a list
of structure fields, and outputs an Emmett file that defines requested
structures and fields within the structures. A user can request all fields in a
structure, or select just a handful of them. The generated file will use the
proper offsets to make an Emmett sparse structure.

The Windows 7 preload uses the output of pdb2structs.py to navigate structures
in a Windows 7 guest and expose certain information.